requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.7.0
pymongo==4.6.1
mistralai==0.4.2
numpy==1.26.4

fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
reportlab==4.0.9